fix: bf16 MoE index_add dtype + add pretrain_sanity Modal smoke test
A 100-step Modal smoke test of the pretrain path surfaced a real bug
that the local fp32 tests couldn't, plus confirmed a known limitation.

FIX (real bug) — MoE dispatch crashed under bf16 autocast:
  model.py:726  RuntimeError: index_add_(): self (BFloat16) and source
  (Float) must have the same scalar type.
  moe_out is bf16 (zeros_like of the bf16 hidden) but the router gate
  comes from the fp32 softmax, so (expert_output * gates) promotes to
  fp32. Fixed both dispatch sites (learned top-2 path + hash-routing
  path) to apply the gate in fp32 (precision) then cast to the output
  dtype before index_add_. Local fp32 tests never exercised autocast,
  so this only appeared on Modal (GPU, bf16).

  Regression: tests/test_model.py adds two CPU-bf16-autocast tests
  (test_bf16_autocast_forward_backward, test_bf16_autocast_hash_routing).
  CPU autocast triggers the same device-agnostic dtype check, so the
  crash is now caught locally for free. 144 -> 146 tests.

train.py — compile_enabled opt-out for run_training:
  Mirrors run_pretrain_extend (defaults True, production unchanged).
  Lets a smoke test run eager so step events appear immediately instead
  of ~10 min of silent torch.compile tracing.

app.py — pretrain_sanity() Modal entry point:
  100-step eager smoke test on the canonical OSRT-605M preset, no
  ckpt/eval/wandb. Footprint reduced (batch 8->1, seq 2048->1024) so it
  fits one H100 in eager: the full config OOMs because the 8 full-vocab
  fp32 logit tensors (1 main + 5 aux-loop + 2 MTP) + the mHC 4x residual
  stream dominate activation memory — the known activation-memory item
  the architecture-optimization review flags (fused linear-CE / grad
  checkpointing are the GPU-phase fixes).

Smoke-test result (app ap-RSl0LF1nNf3jnftpafZ3S2): 100 steps complete,
task loss 10.71 -> ~6.3 (init = ln(32768) = 10.4), MoE balanced
(bal ~1.0, drop 0, emax/emin ~0.14/0.11 — no collapse), ~1,755 tok/s,
22.3 GB. The full v6 stack trains.

NOTE: run_training always saves a FINAL checkpoint regardless of
ckpt_interval, so the sanity wrote a throwaway osrt_v5_final.pt to the
checkpoints volume — worth gating the final save on a sanity flag, or
using a distinct stage prefix, in a follow-up.

+++ src/osrt/model.py
@@ -420,7 +420,12 @@ class MoELayer(nn.Module):
             token_indices = (assign == ei).nonzero(as_tuple=True)[0]
             if token_indices.numel() == 0:
                 continue
-            moe_out.index_add_(0, token_indices, expert(x_flat[token_indices]))
+            # Cast to moe_out.dtype: under bf16 autocast the expert may
+            # emit a dtype that index_add_ rejects against the bf16 buffer.
+            moe_out.index_add_(
+                0, token_indices,
+                expert(x_flat[token_indices]).to(moe_out.dtype),
+            )
         moe_out = moe_out.view(B, S, D)
 
         # Aux losses: deterministic routing has nothing to balance. Keep them as
@@ -723,7 +728,14 @@ class MoELayer(nn.Module):
 
             # Scatter-add: a token may contribute from multiple experts
             # (different ranks), so use index_add, not direct assignment.
-            moe_out.index_add_(0, token_indices, expert_output * gates)
+            # The gate (from the fp32 router softmax) is applied in fp32 for
+            # precision, then cast back to the output dtype — without the
+            # cast, index_add_ raises under bf16 autocast because moe_out is
+            # bf16 while (expert_output * gates) promotes to fp32.
+            moe_out.index_add_(
+                0, token_indices,
+                (expert_output * gates).to(moe_out.dtype),
+            )
 
         moe_out = moe_out.view(B, S, D)
